Object-based overhaul.
My first go round functioned, but as more complex features came into
play (recursion, die meta-data), it quickly became obvious I needed to
reign it in a bit.  This commit replaces most of the public functions
with a single function that can take either an options object, or a Roll
config.